ProgettoDSBD/data_retrieval/requirements.txt
flask
kafka-python
prometheus_api_client
confluent_kafka
prometheus_client
mysql.connector

